requirements/requirements_for_paddlespeech.txt
numpy==1.18
pandas==1.3.5
tokenizers==0.10.1
matplotlib
requests
scipy
wheel
accelerate
argparse
datasets
seqeval
sklearn
torch
tqdm
transformers
typing
paddlenlp
paddlepaddle-gpu
soundfile
librosa==0.8.1
jiwer
torchaudio
paddlehub
pypinyin
LAC
sox
pyworld
paddleaudio
paddlespeech
sentencepiece
adapter-transformers
py7zr
conllu
apache_beam
timm
spacy
ftfy
yacs
importlib_metadata
apache_beam
timm
